fix offload wrappers to only include backend when actually on device
